Update from "publish and package" action

+++ index.js
@@ -400,13 +400,13 @@ async function deletePackageVersion(token, clientId, versionId) {
     packageVersionId: versionId,
     headers: {
       accept: "application/vnd.github.package-deletes-preview+json",
-      authorization: `token ${token}`
-    }
+      authorization: `token ${token}`,
+    },
   })
-    .then(data => {
+    .then((data) => {
       return { versionId, data };
     })
-    .catch(error => {
+    .catch((error) => {
       return { versionId, error };
     });
 }
@@ -425,8 +425,8 @@ async function getRepoPackages(
     maxVersions,
     headers: {
       accept: "application/vnd.github.packages-preview+json",
-      authorization: `token ${token}`
-    }
+      authorization: `token ${token}`,
+    },
   });
 }
 
@@ -448,20 +448,20 @@ if (!owner || !repoName) {
 }
 
 const clientId = "stripethree/gpr-janitor";
-const dryRun = true === core.getInput("dry-run");
-const maxPackagesToFetch = core.getInput("packages-to-fetch");
-const maxVersionsToFetch = core.getInput("versions-to-fetch");
-const minAgeDays = core.getInput("min-age-days");
-const minVersionsToKeep = core.getInput("keep-versions");
+const dryRun = core.getInput("dry-run") === "true";
+const maxPackagesToFetch = parseInt(core.getInput("packages-to-fetch"));
+const maxVersionsToFetch = parseInt(core.getInput("versions-to-fetch"));
+const minAgeDays = parseInt(core.getInput("min-age-days"));
+const minVersionsToKeep = parseInt(core.getInput("keep-versions"));
 
 getRepoPackages(token, owner, repoName, maxPackagesToFetch, maxVersionsToFetch)
-  .then(data => {
+  .then((data) => {
     const packages = data.repository.packages;
 
     const versionsToDelete = [];
 
     // for each package...
-    packages.nodes.forEach(pkg => {
+    packages.nodes.forEach((pkg) => {
       const currentTime = new Date().getTime();
       const msPerDay = 1000 * 60 * 60 * 24;
 
@@ -476,11 +476,11 @@ getRepoPackages(token, owner, repoName, maxPackagesToFetch, maxVersionsToFetch)
 
       const oldVersions = pkgVersions
         .slice(minVersionsToKeep)
-        .filter(pkgVersion => {
+        .filter((pkgVersion) => {
           const pkgUpdatedTime = new Date(
             Math.max.apply(
               null,
-              pkgVersion.files.nodes.map(function(e) {
+              pkgVersion.files.nodes.map(function (e) {
                 return new Date(e.updatedAt);
               })
             )
@@ -489,11 +489,11 @@ getRepoPackages(token, owner, repoName, maxPackagesToFetch, maxVersionsToFetch)
             ((currentTime - pkgUpdatedTime) / msPerDay).toFixed(2) > minAgeDays
           );
         })
-        .filter(pkgVersion => {
+        .filter((pkgVersion) => {
           // never delete the current version
           return pkgVersion.version !== latestVersion;
         })
-        .forEach(pkgVersion => {
+        .forEach((pkgVersion) => {
           console.log(
             `Version ${pkgVersion.version} of package '${owner}/${pkgName}' (id: ${pkgVersion.id}) marked for deletion.`
           );
@@ -506,14 +506,19 @@ getRepoPackages(token, owner, repoName, maxPackagesToFetch, maxVersionsToFetch)
       return [];
     }
 
+    if (!versionsToDelete.length) {
+      console.log("There are no package versions to delete.");
+      return [];
+    }
+
     return Promise.all(
-      versionsToDelete.map(versionId =>
+      versionsToDelete.map((versionId) =>
         deletePackageVersion(token, clientId, versionId)
       )
     );
   })
-  .then(deletions => {
-    outputs = deletions.map(item => {
+  .then((deletions) => {
+    outputs = deletions.map((item) => {
       if (item.data && item.data.deletePackageVersion.success === true) {
         return `Version id ${item.versionId} deleted.`;
       }
@@ -523,6 +528,9 @@ getRepoPackages(token, owner, repoName, maxPackagesToFetch, maxVersionsToFetch)
       return `Unexpected result for version id ${item.versionId}. Details: ${item.data}`;
     });
     console.log(outputs.join("\n"));
+  })
+  .catch((err) => {
+    console.log(err);
   });
